+++ src/config/settings.py
@@ -3,9 +3,8 @@ from datetime import date, timedelta
 
 import yaml
 from dotenv import load_dotenv
-from sqlalchemy import VARCHAR, Integer
 
-from src.utils.general_utils import get_most_recent_date
+from src.utils.general_utils import get_most_recent_date, parse_extra_dims
 
 load_dotenv()
 
@@ -25,19 +24,6 @@ def load_pipeline_config(pipeline_name):
     with open(config_path, "r") as config_file:
         config = yaml.safe_load(config_file)
     return config
-
-
-# TODO shift this to some utils?
-def parse_extra_dims(extra_dims):
-    parsed_extra_dims = {}
-    for extra_dim in extra_dims:
-        dim = next(iter(extra_dim))
-        if extra_dim[dim] == "str":
-            parsed_extra_dims[dim] = VARCHAR
-        else:
-            parsed_extra_dims[dim] = Integer
-
-    return parsed_extra_dims
 
 
 def parse_pipeline_config(dataset, test, update, mode):
