+++ src/utils/cloud_utils.py
@@ -24,21 +24,13 @@ def get_container_client(mode, container_name):
     azure.storage.blob.ContainerClient
         A `ContainerClient` object that can be used to interact with the specified Azure Blob Storage container
 
-
+    """
     blob_sas = os.getenv(f"DSCI_AZ_SAS_{mode.upper()}")
     blob_url = (
         f"https://imb0chd0{mode}.blob.core.windows.net/"
         + container_name  # noqa
         + "?"  # noqa
         + blob_sas  # noqa
-    )
-    """
-    blob_sas = os.getenv(f"DSCI_AZ_SAS_{mode.upper()}")
-    blob_url = (
-            f"https://imb0chd0{mode}.blob.core.windows.net/"
-            + container_name  # noqa
-            + "?"  # noqa
-            + blob_sas  # noqa
     )
     return ContainerClient.from_container_url(blob_url)
